Сделал jsDocs пояснения для функций общения с сервером !!Сделал 2 уровня, для теста экрана выигрыша (а то эти 10 уровней игру фиг пройдёшь просто так:)
