+++ js/data/constants.js
@@ -1,11 +1,17 @@
-export const LEVELS_AMOUNT = 3;
+export const LEVELS_AMOUNT = 2;
 export const LIVES_AMOUNT = 3;
 export const TIME_AMOUNT = 5 * 60;
 export const FAST_TIME = 30;
-export const LEVEL_TYPE = {
-  GENRE: `genre`,
-  ARTIST: `artist`
-};
-export const SERVER_URL = `https://es.dump.academy/guess-melody`;
 
+export const SERVER_URL = `https://es.dump.academy/guess-melody`;
 export const DEFAULT_NAME = `electrovlad_id123103`;
+export const OTHER_RESULTS_MOCK = [{
+  time: 10,
+  points: -1
+}, {
+  time: 20,
+  points: 2
+}, {
+  time: 30,
+  points: 3
+}];
