+++ js/main.js
@@ -1,5 +1,5 @@
 import App from './app';
-import {loadData} from './functions/loadFunctions';
+import {loadData} from './functions/serverCommunicationFunctions';
 
 loadData().then((levelsSet) => App.init(levelsSet));
 App.showWelcome();
